auth-server-requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.1
python-dotenv==1.0.0
